feat: add itemList component
