+++ components/timeline/index.zh-CN.md
@@ -18,16 +18,14 @@ demo:
 ## 代码演示
 
 <!-- prettier-ignore -->
-[//]: # (<code src="./demo/basic.tsx">基本用法</code>)
-[//]: # '<code src="./demo/color.tsx">圆圈颜色</code>'
-
+<code src="./demo/basic.tsx">基本用法</code>
+<code src="./demo/color.tsx">圆圈颜色</code>
 <code src="./demo/pending.tsx">最后一个及排序</code>
-
-[//]: # '<code src="./demo/alternate.tsx">交替展现</code>'
-[//]: # '<code src="./demo/custom.tsx">自定义时间轴点</code>'
-[//]: # '<code src="./demo/right.tsx">右侧时间轴点</code>'
-[//]: # '<code src="./demo/label.tsx">标签</code>'
-[//]: # '<code src="./demo/wireframe.tsx" debug>线框风格</code>'
+<code src="./demo/alternate.tsx">交替展现</code>
+<code src="./demo/custom.tsx">自定义时间轴点</code>
+<code src="./demo/right.tsx">右侧时间轴点</code>
+<code src="./demo/label.tsx">标签</code>
+<code src="./demo/wireframe.tsx" debug>线框风格</code>
 
 ```__react
 import Alert from '../alert';
